Set up fixtures for mouse actions in wdspec tests; r=ato+446296
MozReview-Commit-ID: zzIQgSgEPb

Upstreamed from https://bugzilla.mozilla.org/show_bug.cgi?id=1337133
