+++ webdriver/actions/mouse.py
@@ -1,4 +1,8 @@
-import pytest
+from support.refine import get_events
 
-def test_nothing():
-    assert True
+
+#def test_nothing(session, test_actions_page, mouse_chain):
+#    mouse_chain \
+#        .pointer_down(1) \
+#        .perform()
+#    assert True
